Bbe add!!!

+++ src/TowerDefense/Entity/Enemy/BossEnemy.java
@@ -9,7 +9,7 @@ import java.io.FileInputStream;
 public class BossEnemy extends Enemy {
     public static final String IMAGE_URL = "file:img/PNG/Retina/towerDefense_tile246.png";
 
-    public BossEnemy(int x, int y) {
+    public BossEnemy(double x, double y) {
         super(x, y, 20, 2.5);
         img = new Image(IMAGE_URL);
         setDir('E');
