+++ src/TowerDefense/Entity/Enemy/SmallerEnemy.java
@@ -9,7 +9,7 @@ import java.io.FileInputStream;
 public class SmallerEnemy extends Enemy {
     public static final String IMAGE_URL = "file:img/PNG/Retina/towerDefense_tile248.png";
 
-    public SmallerEnemy(int x, int y) {
+    public SmallerEnemy(double x, double y) {
         super(x, y, 5, 3);
         img = new Image(IMAGE_URL);
         setDir('E');
